Refactor: authReducer in separate file.
