+++ src/context/auth-context.js
@@ -1,4 +1,5 @@
 import { createContext, useContext, useReducer } from "react";
+import { authReducer } from "../redux/index";
 
 const defaultValue = { user: "", token: "" };
 
@@ -6,23 +7,7 @@ const AuthContext = createContext(defaultValue);
 
 const AuthProvider = ({ children }) => {
 
-    const initialAuthState = {
-        user: "",
-        token: ""
-    }
-
-    const authReducer = (state, action) => {
-        switch (action.type) {
-            case "SIGN_UP":
-                return { user: action.payload.user, token: action.payload.token };
-            case "LOGIN":
-                return { user: action.payload.user, token: action.payload.token };
-            case "LOGOUT":
-                return initialAuthState;
-            default:
-                return state;
-        }
-    }
+    const initialAuthState = { user: "", token: "" };
 
     const [authState, authDispatch] = useReducer(authReducer, initialAuthState);
 
@@ -30,7 +15,7 @@ const AuthProvider = ({ children }) => {
         <AuthContext.Provider value={{ authDispatch, authState }} >
             {children}
         </AuthContext.Provider>
-    )
+    );
 }
 
 const useAuth = () => useContext(AuthContext);
