+++ src/redux/index.js
@@ -1,0 +1,1 @@
+export { authReducer } from "./authReducer";
